frontend/requirements.txt
# Streamlit
streamlit==1.30.0

# WebSocket
websocket-client==1.7.0

# HTTP Requests
requests==2.31.0

